rag/requirements.txt
qdrant-client==1.14.2
openai>=1.51.0
python-dotenv==1.0.1
